# [core][misc] improve free_finished_seq_groups