@youkaichao 2024-07-27
Here is the flamechart of the main branch, with 100k prompts. It becomes clear that `free_finished_seq_groups` is the major source of overhead.

![flamechart](https://github.com/user-attachments/assets/adae061d-c70b-4b27-85cc-c6520b5c8e54)