@youkaichao 2024-07-27
Here is the flame chart of this PR, with various number of prompts:

10k prompts:

![flamechart_10k](https://github.com/user-attachments/assets/84fce574-41a7-4057-8e24-71c253dd1560)

100k prompts:

![flamechart_100k](https://github.com/user-attachments/assets/7eca542c-bc83-4a4e-a374-7612dd10d10a)

500k prompts:
![flamechart_500k](https://github.com/user-attachments/assets/de7c61a0-fdc6-4353-b060-e880e758e05c)

Even after this PR, there is still a slow down factor that is proportional to the number of waiting requests. The flamechat points to:

https://github.com/vllm-project/vllm/blob/fad5576c58864a6c2cf528f67e60e03a949b3dac/vllm/core/scheduler.py#L680 

which is indeed very inefficient.