@youkaichao 2024-07-30
> @youkaichao I think you must rebase to pass the failing Examples Test
> 
> ```
> [2024-07-29T19:01:27Z] python3: can't open file '/vllm-workspace/examples/llava_example.py': [Errno 2] No such file or directory
> ```

If there's no more comments, I will force-merge to avoid one additional ci cost :)